Log in 1 place, error on null value

+++ AutoSiteProject/AutoSiteProject.UI/Controllers/CarBodyTypeController.cs
@@ -5,6 +5,7 @@ using AutoSiteProject.Models.Bl.Interfaces;
 using AutoSiteProject.Models.Bl.Interfaces.FieldCopiers;
 using AutoSiteProject.Models.ViewModels;
 using AutoSiteProject.Models.DB;
+using System;
 
 namespace AutoSiteProject.UI.Controllers
 {
@@ -53,6 +54,7 @@ namespace AutoSiteProject.UI.Controllers
         public ActionResult Edit(int id)
         {
             var dbItem = _carBodyTypeManager.GetById(id);
+            if (dbItem == null) throw new NullReferenceException();
             return View(_carBodyTypeFieldCopier.CopyFields(dbItem,new CarBodyTypeViewModel()));
         }
         //Post
@@ -62,6 +64,7 @@ namespace AutoSiteProject.UI.Controllers
             if (ModelState.IsValid)
             {
                 var dbItem = _carBodyTypeManager.GetById(model.Id);
+                if (dbItem == null) throw new NullReferenceException();
                 dbItem = _carBodyTypeFieldCopier.CopyFields(model, dbItem);
                 _carBodyTypeManager.Edit(dbItem);
                 return RedirectToAction("List");
@@ -73,6 +76,7 @@ namespace AutoSiteProject.UI.Controllers
         public ActionResult Delete(int id)
         {
             var dbItem = _carBodyTypeManager.GetById(id);
+            if (dbItem == null) throw new NullReferenceException();
             _carBodyTypeManager.Delete(dbItem);
             return RedirectToAction("List");
         }
